Remove migrations! Yolo!
