--- src/app/store/actions/album.actions.ts
@@ -1,9 +1,0 @@
-export class SaveAlbum {
-  static readonly type = '[Album] SaveAlbum';
-  constructor(public id: string) {}
-}
-
-export class RemoveAlbum {
-  static readonly type = '[Album] RemoveAlbum';
-  constructor(public id: string) {}
-}
