--- src/app/store/actions/artist.actions.ts
@@ -1,9 +1,0 @@
-export class SaveArtist {
-  static readonly type = '[Artist] SaveArtist';
-  constructor(public id: string) {}
-}
-
-export class RemoveArtist {
-  static readonly type = '[Artist] RemoveArtist';
-  constructor(public id: string) {}
-}
