--- src/app/store/models/album-state.model.ts
@@ -1,3 +1,0 @@
-export interface AlbumStateModel {
-  ids: string[];
-}
