--- src/app/store/models/artist-state.model.ts
@@ -1,3 +1,0 @@
-export interface ArtistStateModel {
-  ids: string[];
-}
